# bin/Release/net6.0/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "atm, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null",
    "version": "1.0"
  },
  "paths": {
    "/api/Atm": {
      "get": {
        "tags": [
          "Atm"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "string"
            }
          }
        }
      }
    },
    "/test": {
      "get": {
        "tags": [
          "Atm"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "string"
            }
          }
        }
      }
    },
    "/register": {
      "post": {
        "tags": [
          "Atm"
        ],
        "consumes": [
          "multipart/form-data"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "formData",
            "name": "CustomerID",
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "formData",
            "name": "ID",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "formData",
            "name": "Email",
            "required": true,
            "type": "string"
          },
          {
            "in": "formData",
            "name": "KRAPIN",
            "required": true,
            "type": "string"
          },
          {
            "in": "formData",
            "name": "AccountName",
            "required": true,
            "type": "string"
          },
          {
            "in": "formData",
            "name": "AccountNumber",
            "required": true,
            "type": "string"
          },
          {
            "in": "formData",
            "name": "PIN",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "formData",
            "name": "MobileNumber",
            "required": true,
            "type": "string"
          },
          {
            "in": "formData",
            "name": "Limit",
            "required": true,
            "type": "number",
            "format": "float"
          },
          {
            "in": "formData",
            "name": "AvailableBalance",
            "required": true,
            "type": "number",
            "format": "float"
          },
          {
            "in": "formData",
            "name": "ActualBalance",
            "required": true,
            "type": "number",
            "format": "float"
          },
          {
            "in": "formData",
            "name": "RegistrationDate",
            "required": true,
            "type": "string",
            "format": "date-time"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "string"
            }
          }
        }
      }
    },
    "/login": {
      "post": {
        "tags": [
          "Atm"
        ],
        "consumes": [
          "multipart/form-data"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "formData",
            "name": "Email",
            "required": true,
            "type": "string"
          },
          {
            "in": "formData",
            "name": "PIN",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "string"
            }
          }
        }
      }
    },
    "/withdraw": {
      "post": {
        "tags": [
          "Atm"
        ],
        "consumes": [
          "multipart/form-data"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "formData",
            "name": "WithdrawalID",
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "formData",
            "name": "TransactionID",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "formData",
            "name": "CustomerID",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "formData",
            "name": "Amount",
            "required": true,
            "type": "number",
            "format": "float"
          },
          {
            "in": "formData",
            "name": "WithdrawalDate",
            "required": true,
            "type": "string",
            "format": "date-time"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "string"
            }
          }
        }
      }
    },
    "/balance": {
      "post": {
        "tags": [
          "Atm"
        ],
        "consumes": [
          "multipart/form-data"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "formData",
            "name": "Email",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "string"
            }
          }
        }
      }
    },
    "/transfer": {
      "post": {
        "tags": [
          "Atm"
        ],
        "consumes": [
          "multipart/form-data"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "formData",
            "name": "TransferID",
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "formData",
            "name": "TransactionID",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "formData",
            "name": "SenderEmail",
            "required": true,
            "type": "string"
          },
          {
            "in": "formData",
            "name": "RecipientAccount",
            "required": true,
            "type": "string"
          },
          {
            "in": "formData",
            "name": "Amount",
            "required": true,
            "type": "number",
            "format": "float"
          },
          {
            "in": "formData",
            "name": "TransferDate",
            "required": true,
            "type": "string",
            "format": "date-time"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "string"
            }
          }
        }
      }
    }
  }
}